chore: Shuffle around things for shared consideration
